+++ c++/pod5_format_pybind/bindings.cpp
@@ -16,10 +16,7 @@ PYBIND11_MODULE(pod5_format_pybind, m)
         .export_values();
 
     py::class_<FileWriterOptions>(m, "FileWriterOptions")
-        .def(py::init([]() {
-            FileWriterOptions options;
-            return options;
-        }))
+        .def(py::init<>())
         .def_property(
             "max_signal_chunk_size",
             &FileWriterOptions::max_signal_chunk_size,
@@ -95,7 +92,32 @@ PYBIND11_MODULE(pod5_format_pybind, m)
 
     // Opening files
     m.def("open_file", &open_file, "Open a POD5 file for reading");
-    m.def("recover_file", &recover_file, "Recover a POD5 file which was not closed correctly");
+
+    // Recovering files
+    py::class_<RecoverFileOptions>(m, "RecoverFileOptions")
+        .def(py::init<>())
+        .def_readwrite("file_writer_options", &RecoverFileOptions::file_writer_options)
+        .def_readwrite("cleanup", &RecoverFileOptions::cleanup);
+    py::class_<RecoveredRowCounts>(m, "RecoveredRowCounts")
+        .def(py::init<>())
+        .def_readwrite("signal", &RecoveredRowCounts::signal)
+        .def_readwrite("run_info", &RecoveredRowCounts::run_info)
+        .def_readwrite("reads", &RecoveredRowCounts::reads);
+    py::class_<CleanupError>(m, "CleanupError")
+        .def(py::init<>())
+        .def_readwrite("file_path", &CleanupError::file_path)
+        .def_readwrite("description", &CleanupError::description);
+    py::class_<RecoveryDetails>(m, "RecoveryDetails")
+        .def(py::init<>())
+        .def_readwrite("row_counts", &RecoveryDetails::row_counts)
+        .def_readwrite("cleanup_errors", &RecoveryDetails::cleanup_errors);
+    m.def(
+        "recover_file",
+        &::recover_file,
+        "Recover a POD5 file which was not closed correctly",
+        py::arg("src_filename"),
+        py::arg("dest_filename"),
+        py::arg("options") = nullptr);
 
     m.def(
         "update_file",